Run the script as regular user
I dislike running scripts like this as root. The script should ask for root
permission for the operations that need them and execute everything else as
regular user. We are pulling things from the web and then running cmake on
them, which allows execution of random code... so don't do that as root.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>