Allow running generate_* with set -e
mkdir's exit code won't be 0 if the directory already exists, regardless of
whether -p parameter is used or not.

When a build script calls `set -e` at the beginning to automatically stop a
build when a command fails, the generate_* helpers will fail due to mkdir
being likely to fail. This commit makes sure the exit code is ignored for
the mkdir calls.