Add post script section to ingredient
Post script gives the recipe a chance to modify the downloaded packages
from ingredient before installation.

One common use case is for a user to build an AppImage based on a
custom-built package from the offical source. The recipe can rely on
apt to pull dependency from the offical source. Once all packages are
downloaded, the recipe can use post_script to replace the offical
package with the custom-built one.