Exclude: libfontconfig1-dev
I was just about to create FreeCAD related PR. To remove FreeType from the FreeCAD recipe and have seen this:

https://github.com/AppImage/AppImages/pull/323

I can confirm we where affected by the mentioned issue too. In addition we use libfontconfig1-dev package as a workaround on PPA (for not having to create a symlink). I guess it is harmless if it is added to the AppImage. But as it doesn't do anything from the AppImage perspective. Likely best to remove it altogether.

https://forum.freecadweb.org/viewtopic.php?f=10&t=15525&start=190#p222348