Remove libthai from the excludelist and improve excludelist formatting (#555)
* Removed libthai from the excludelist and improved excludelist formatting

libthai.so.0 has been removed from the excludelist.
The original reason why libthai had been added to it is an issue that occurs when libpango is not bundled but libthai is (see #538). But since then, libpango had been removed from the excludelist in 2022. This means that the original reason why libthai had been excluded is no longer relevant. Therefore it could be removed from the excludelist. This fixes #538.

Additionally, the excludelist formatting has been improved. There had been different places scattered across the excludelist in which libraries had been excluded or not excluded due to other libraries (like libthai / libpango). They have all been brought together (with all information and reasoning) in one block to prevent issues like this in the future.
Additionally, missing information out of #528 has been added to this block.

* uses: actions/checkout@v4, since v2 does not work anymore

* Need image: ubuntu:20.04 so that checkout can work?

---------

Co-authored-by: probonopd <probonopd@users.noreply.github.com>